adgm_corporate_agent/requirements.txt
streamlit
python-docx
openai
python-dotenv
pypdf
tiktoken
numpy
scipy